# openapi.yaml
openapi: 3.0.3
info:
  title: Cratebase API
  version: "0.1.0"
  description: >
    Dynamic collections, auth, file storage, and realtime over one REST API.
    Error responses always use the shape in `ErrorBody`. Bearer tokens
    (admin or auth-collection) go in `Authorization: Bearer <token>`.
servers:
  - url: /api
security:
  - bearerAuth: []
  - {}
paths:
...
  /collections/{idOrName}/records:
    parameters:
      - name: idOrName
        in: path
        required: true
        schema: { type: string }
    get:
      summary: List/search records
      parameters:
        - name: filter
          in: query
          schema: { type: string }
          description: 'Filter expression, e.g. `published = true && title ~ "hello"`'
        - name: sort
          in: query
          schema: { type: string }
          description: Comma-separated fields, `-` prefix for descending (default `-created`)
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: perPage
          in: query
          schema: { type: integer, default: 30, maximum: 500 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  page: { type: integer }
                  perPage: { type: integer }
                  totalItems: { type: integer }
                  totalPages: { type: integer }
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/Record" }
        "403": { $ref: "#/components/responses/Error" }
    post:
      summary: Create a record (JSON, or multipart/form-data when the collection has file fields)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
          multipart/form-data:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
  /collections/{idOrName}/records/{id}:
    parameters:
      - name: idOrName
        in: path
        required: true
        schema: { type: string }
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      summary: View one record
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update a record (JSON, or multipart/form-data for file fields)
      requestBody:
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
          multipart/form-data:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Record" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a record
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          type: object
          additionalProperties:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
          description: Per-field validation errors, present when code is 400
...
    Record:
      type: object
      description: Every record has these plus its collection's own schema fields.
      properties:
        id: { type: string }
        created: { type: string }
        updated: { type: string }
        collectionId: { type: string }
        collectionName: { type: string }
      additionalProperties: true
...